Add DMR x-axis options and save DMR-parameters
Add xScale options 'minmax', 'factor' and 'outlier' to DMR-plots. The first sets the minimum and maximum x-value, defined in xScaleValue and seperated by a space. The second sets the x-range to the standart devation times the value given in xScaleValue. The last includes all outliers.

Write a TTree, which stores rms, mu and deltamu values of the DMR-plots in different branches for each subdetector, and a second TTree for the y-direction of the pixel detectors.

The TTree simplifies access to the parameters of the DMR-plots, making them easier to use for further validations.